Changed type

+++ test/cpu_fusion.cpp
@@ -2815,7 +2815,7 @@ TEST(cpu_fusion, fuse_bounded_relu_inter_vs_cpu)
 TEST(cpu_fusion, fuse_dropout)
 {
     auto make_function =
-        [](Shape input_shape, const uint32_t seed_val, float one_minus_prob, bool fuse) {
+        [](Shape input_shape, const uint32_t seed_val, double one_minus_prob, bool fuse) {
             auto input = std::make_shared<op::Parameter>(element::f32, input_shape);
             auto value = op::Constant::create(element::f32, input_shape, {one_minus_prob});
             auto const1 = op::Constant::create(input->get_element_type(), Shape{}, {1});
@@ -2824,7 +2824,7 @@ TEST(cpu_fusion, fuse_dropout)
                                                                input->get_shape(),
                                                                input->get_element_type(),
                                                                seed_val,
-                                                               (double)one_minus_prob);
+                                                               one_minus_prob);
 
             auto mult = std::make_shared<op::Multiply>(gen_mask, input);
